+++ src/examples/error_handling/longjmp_with_stack_trace.cc
@@ -56,7 +56,7 @@ inline void error_create(const char *message) {
 	p->symbols=backtrace_symbols(p->array, p->size);
 	strncpy(p->message, message, max_message_size);
 	last_error=p;
-	longjmp(env, (int)p);
+	longjmp(env, (unsigned long)p);
 }
 
 inline void error_print(FILE *f, error_data *p) {
@@ -86,7 +86,7 @@ inline void error_free_last() {
 // This function **must** be inlined as if setjmp returns then env will no longer
 // be valid. That's why we don't use it (there is no way to guarantee inlining).
 inline error_data *error_setjmp() {
-	int ret=setjmp(env);
+	unsigned long ret=setjmp(env);
 	if (!ret) {
 		return(NULL);
 	} else {
@@ -95,7 +95,7 @@ inline error_data *error_setjmp() {
 }
 
 // here is another idea: why not use a macro instead:
-#define mac_error_setjmp() ((error_data *)setjmp(env))
+#define mac_error_setjmp() ((error_data *)(unsigned long)setjmp(env))
 
 /* this simulates a function which sometimes encounters errors */
 void func() {
